requirements.txt
#
# This file is autogenerated by pip-compile with python 3.6
# To update, run:
#
#    pip-compile
#
amqp==5.1.1
    # via kombu
appy==1.0.11
    # via django-appypod
asgiref==3.4.1
    # via django
async-timeout==4.0.2
    # via redis
attrs==21.4.0
    # via fiona
beautifulsoup4==4.11.1
    # via mapentity
billiard==3.6.4.0
    # via celery
btrees==4.10.0
    # via zodb
cached-property==1.5.2
    # via kombu
cairocffi==0.9.0
    # via
    #   cairosvg
    #   geotrek (setup.py)
    #   mapentity
    #   weasyprint
cairosvg==2.5.2
    # via
    #   geotrek (setup.py)
    #   weasyprint
celery[redis]==5.1.2
    # via
    #   django-celery-results
    #   geotrek (setup.py)
certifi==2021.10.8
    # via
    #   fiona
    #   requests
    #   sentry-sdk
cffi==1.15.0
    # via
    #   cairocffi
    #   persistent
    #   weasyprint
charset-normalizer==2.0.12
    # via requests
click==7.1.2
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   cligj
    #   fiona
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via
    #   celery
    #   fiona
click-repl==0.2.0
    # via celery
cligj==0.7.2
    # via fiona
coreapi==2.3.3
    # via drf-yasg
coreschema==0.0.4
    # via
    #   coreapi
    #   drf-yasg
coverage==6.2
    # via geotrek (setup.py)
cssselect2==0.4.1
    # via
    #   cairosvg
    #   weasyprint
datetime==4.4
    # via appy
defusedxml==0.7.1
    # via cairosvg
deprecated==1.2.13
    # via redis
django==3.2.16
    # via
    #   django-appconf
    #   django-appypod
    #   django-clearcache
    #   django-debug-toolbar
    #   django-embed-video
    #   django-extensions
    #   django-filter
    #   django-js-asset
    #   django-leaflet
    #   django-modeltranslation
    #   django-weasyprint
    #   djangorestframework
    #   drf-yasg
    #   easy-thumbnails
    #   geotrek (setup.py)
    #   mapentity
    #   paperclip
django-appconf==1.0.5
    # via django-compressor
django-appypod==2.0.7
    # via mapentity
django-celery-results==2.2.0
    # via geotrek (setup.py)
django-clearcache @ git+https://github.com/timonweb/django-clearcache.git@6de7994
    # via geotrek (setup.py)
django-colorfield==0.6.3
    # via geotrek (setup.py)
django-compressor==3.1
    # via mapentity
django-crispy-forms==1.13.0
    # via mapentity
django-debug-toolbar==3.2.4
    # via geotrek (setup.py)
django-embed-video==1.4.3
    # via
    #   mapentity
    #   paperclip
django-extended-choices==1.3.3
    # via geotrek (setup.py)
django-extensions==3.1.5
    # via geotrek (setup.py)
django-filter==21.1
    # via mapentity
django-js-asset==2.0.0
    # via django-mptt
django-leaflet==0.19.post9
    # via mapentity
django-modelcluster==5.3
    # via geotrek (setup.py)
django-modeltranslation==0.17.5
    # via mapentity
django-mptt==0.13.4
    # via geotrek (setup.py)
django-tinymce==3.4.0
    # via mapentity
django-weasyprint==1.1.0.post2
    # via
    #   geotrek (setup.py)
    #   mapentity
djangorestframework==3.13.1
    # via
    #   djangorestframework-datatables
    #   djangorestframework-gis
    #   drf-extensions
    #   drf-yasg
    #   mapentity
djangorestframework-datatables==0.7.0
    # via mapentity
djangorestframework-gis==0.18
    # via mapentity
docutils==0.18.1
    # via geotrek (setup.py)
drf-dynamic-fields==0.4.0
    # via geotrek (setup.py)
drf-extensions==0.7.1
    # via geotrek (setup.py)
drf-yasg==1.20.0
    # via geotrek (setup.py)
easy-thumbnails==2.7
    # via
    #   mapentity
    #   paperclip
env-file==2020.12.3
    # via geotrek (setup.py)
factory-boy==3.2.1
    # via geotrek (setup.py)
faker==13.3.4
    # via factory-boy
fiona==1.8.21
    # via
    #   geotrek (setup.py)
    #   mapentity
freezegun==1.2.1
    # via geotrek (setup.py)
geojson==2.5.0
    # via
    #   geotrek (setup.py)
    #   tif2geojson
gpxpy==1.5.0
    # via mapentity
gunicorn==20.1.0
    # via geotrek (setup.py)
html5lib==1.1
    # via weasyprint
idna==3.3
    # via requests
importlib-metadata==4.8.3
    # via
    #   kombu
    #   markdown
    #   redis
inflection==0.5.1
    # via drf-yasg
itypes==1.2.0
    # via coreapi
jinja2==3.0.3
    # via coreschema
kombu==5.1.0
    # via celery
landez==2.5.0
    # via geotrek (setup.py)
lxml==4.9.1
    # via mapentity
mapentity==8.2.1
    # via geotrek (setup.py)
markdown==3.3.7
    # via geotrek (setup.py)
markupsafe==2.0.1
    # via jinja2
mbutil==0.3.0
    # via landez
munch==2.5.0
    # via fiona
netifaces==0.11.0
    # via mapentity
packaging==21.3
    # via
    #   drf-yasg
    #   redis
paperclip==2.6.1
    # via
    #   geotrek (setup.py)
    #   mapentity
papersize==1.0.1
    # via pdfimpose
pdfimpose==1.1.0
    # via geotrek (setup.py)
persistent==4.9.0
    # via
    #   btrees
    #   zodb
pillow==8.4.0
    # via
    #   cairosvg
    #   django-colorfield
    #   easy-thumbnails
    #   geotrek (setup.py)
    #   paperclip
    #   weasyprint
prompt-toolkit==3.0.30
    # via click-repl
psycopg2==2.9.3
    # via geotrek (setup.py)
pycparser==2.21
    # via cffi
pygal==3.0.0
    # via geotrek (setup.py)
pymemcache==3.5.2
    # via geotrek (setup.py)
pyparsing==3.0.8
    # via packaging
pypdf2==1.27.5
    # via pdfimpose
pyphen==0.11.0
    # via weasyprint
python-dateutil==2.8.2
    # via
    #   faker
    #   freezegun
pytz==2022.1
    # via
    #   celery
    #   datetime
    #   django
    #   django-modelcluster
    #   djangorestframework
    #   djangorestframework-datatables
rcssmin==1.1.0
    # via django-compressor
redis==4.2.2
    # via celery
requests==2.27.1
    # via
    #   coreapi
    #   django-embed-video
    #   landez
    #   mapentity
rjsmin==1.2.0
    # via django-compressor
ruamel.yaml==0.17.21
    # via drf-yasg
ruamel.yaml.clib==0.2.6
    # via ruamel.yaml
sentry-sdk==1.5.9
    # via geotrek (setup.py)
simplekml==1.3.6
    # via geotrek (setup.py)
six==1.16.0
    # via
    #   click-repl
    #   django-extended-choices
    #   django-modeltranslation
    #   fiona
    #   html5lib
    #   munch
    #   pymemcache
    #   python-dateutil
    #   zodb
soupsieve==2.3.2
    # via beautifulsoup4
sqlparse==0.4.2
    # via
    #   django
    #   django-debug-toolbar
tif2geojson==0.1.3
    # via geotrek (setup.py)
tinycss2==1.1.1
    # via
    #   cairosvg
    #   cssselect2
    #   weasyprint
transaction==3.0.1
    # via zodb
typing-extensions==4.1.1
    # via
    #   asgiref
    #   async-timeout
    #   faker
    #   importlib-metadata
    #   redis
uritemplate==4.1.1
    # via
    #   coreapi
    #   drf-yasg
urllib3==1.26.9
    # via
    #   requests
    #   sentry-sdk
values==2020.12.3
    # via env-file
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.5
    # via prompt-toolkit
weasyprint==52.5
    # via
    #   django-weasyprint
    #   geotrek (setup.py)
    #   mapentity
webencodings==0.5.1
    # via
    #   cssselect2
    #   html5lib
    #   tinycss2
wrapt==1.14.0
    # via deprecated
xlrd==2.0.1
    # via geotrek (setup.py)
xmltodict==0.12.0
    # via tif2geojson
zc.lockfile==2.0
    # via zodb
zconfig==3.6.0
    # via zodb
zipp==3.6.0
    # via importlib-metadata
zodb==5.7.0
    # via appy
zodbpickle==2.2.0
    # via zodb
zope.interface==5.4.0
    # via
    #   btrees
    #   datetime
    #   persistent
    #   transaction
    #   zodb
pyopenair==1.1.0
# The following packages are considered to be unsafe in a requirements file:
# setuptools
